feat : confirmation box before account deletion

+++ Controller/ProfileController.php
@@ -21,13 +21,20 @@ class ProfileController extends Controller
             echo '<script>window.location.href = "index.php";</script>';
         }
         elseif (isset($_POST['suppression'])) {
-            $this->manager->deleteUser();
-            $dir = $_SESSION['dir'];
-            session_destroy();
-            session_start();
-            $_SESSION['dir'] = $dir;
-            $_GET['action'] = '';
-            echo '<script>window.location.href = "index.php";</script>';
+            ?>
+            <script>
+                document.cookie = "confirm=" + escape(confirm("Confirmez la suppression du compte.")) + "; path=/";
+            </script>
+            <?php
+            if($_COOKIE['confirm'] == 'true') {
+                $this->manager->deleteUser();
+                $dir = $_SESSION['dir'];
+                session_destroy();
+                session_start();
+                $_SESSION['dir'] = $dir;
+                $_GET['action'] = '';
+                echo '<script>window.location.href = "index.php";</script>';
+            }
         }
         include $_SESSION['dir'] . '/View/ProfileView.php';
     }
